flask/requirements.txt
psycopg2-binary
flask
json5
requests
flask-cors
environ
python-dotenv